requirements.txt
ConfigUpdater==3.2
json5==0.13.0
matplotlib==3.10.1
pandas==2.2.3
Pillow==11.1.0
platformdirs==4.9.2
PyMuPDF==1.27.1
PySide6==6.8.2.1
